refactor(computer-use): give the helper server a layer per responsibility
mod.rs had grown to hold the whole server: accepting a connection, framing a
protocol version check, routing every method, asking the host for approval,
naming and launching applications, and the shared model an observation is made
of. Eight hundred lines where a reader has to hold all six concerns at once,
and where the next change lands wherever there is room.

The concerns become files. connection accepts and serves in order; dispatch
routes one request past the shared guards; approval asks the host; app_identity
names and starts an application; state holds what the platform leaves share.
mod.rs keeps only the module wiring and the two protocol facts everything needs.

The boundary was not chosen by eye. Every `use super::super::` in a platform
leaf resolved to an item that is now in state.rs -- that set is the definition
of what the two platforms share, so it is what the module contains. The leaves
now name state directly rather than a re-export, which keeps the dependency
readable and survives the next move.

Item names before and after the carve match exactly: thirty in, thirty out,
none lost and none invented. Windows checks clean and its tests pass.
