+++ console/src-tauri/src/computer_use_server/platform_macos/window.rs
@@ -16,7 +16,7 @@ use core_graphics::window::{
 use serde_json::{json, Value};
 use std::path::{Path, PathBuf};
 
-use super::super::{merge_app_list, InstalledApp, WindowInfo};
+use super::super::state::{merge_app_list, InstalledApp, WindowInfo};
 use super::accessibility_tree::find_ax_window;
 use super::input::reject_recent_user_intervention;
 use super::{dict_i64, dict_string, window_owner_pid};
