+++ console/src-tauri/src/computer_use_server/state.rs
@@ -1,0 +1,327 @@
+//! What the platform leaves share: the shape of a window, an observation, and
+//! the per-connection store that ties an action back to what was observed.
+//!
+//! Every `use super::super::` in a platform leaf resolves here, which is the
+//! boundary this module is drawn along: nothing in it touches an OS API, and
+//! everything in it has to mean the same thing on both platforms. The limits
+//! that shape an observation live here for the same reason -- a screenshot
+//! delivered at different sizes per platform, or document text truncated at
+//! different lengths, would make the observation contract platform-dependent.
+
+use serde_json::{json, Value};
+use std::collections::HashMap;
+use std::sync::atomic::{AtomicU64, Ordering};
+
+/// Native accessibility element handle stored in an accessibility snapshot.
+/// Windows uses a UI Automation element; macOS uses an AXUIElement wrapper.
+#[cfg(windows)]
+pub(super) type NativeElement = windows::Win32::UI::Accessibility::IUIAutomationElement;
+#[cfg(target_os = "macos")]
+pub(super) type NativeElement = super::platform_macos::AxElement;
+
+/// How recently a person must have used the keyboard or mouse for an action to
+/// be refused as racing them.
+pub(super) const USER_INTERVENTION_GRACE_MS: u32 = 750;
+
+// Raw window captures are 32bpp bitmaps; re-encode them as JPEG so a
+// single screenshot costs hundreds of kilobytes instead of tens of
+// megabytes once it is base64-encoded into the response payload.
+pub(super) const SCREENSHOT_JPEG_QUALITY: f32 = 0.8;
+
+// Cap the longest edge of a delivered screenshot. High-resolution
+// displays (for example 4K) would otherwise produce multi-megabyte
+// base64 payloads that inflate the response and the model's image cost.
+// Downscaling to a bounded edge keeps the payload small while leaving
+// enough detail for reading on-screen text and controls.
+pub(super) const SCREENSHOT_MAX_EDGE: u32 = 1600;
+
+// Only the Windows capture path decodes raw bitmaps.
+#[cfg(windows)]
+pub(super) const BMP_HEADER_BYTES: usize = 54;
+
+/// Upper bound on the document text handed back with an observation. A large
+/// document would otherwise dominate the model's context, and the leading
+/// portion is what identifies the current state.
+pub(super) const DOC_TEXT_MAX: usize = 4000;
+
+static NEXT_ID: AtomicU64 = AtomicU64::new(1);
+
+#[derive(Clone)]
+pub(super) struct WindowInfo {
+    pub(super) hwnd: isize,
+    pub(super) app_id: String,
+    pub(super) display_name: String,
+    pub(super) title: String,
+    // Windows matches this against its credential-dialog guard. macOS has no
+    // equivalent notion of a window class and recognises those dialogs by
+    // title and owner instead.
+    #[cfg_attr(target_os = "macos", allow(dead_code))]
+    pub(super) class_name: String,
+}
+
+impl WindowInfo {
+    pub(super) fn to_json(&self) -> Value {
+        json!({
+            "app_id": self.app_id,
+            "id": self.hwnd.to_string(),
+            "title": self.title,
+        })
+    }
+}
+
+pub(super) struct Snapshot {
+    pub(super) window: WindowInfo,
+    /// The window's on-screen rectangle as `[left, top, width, height]`.
+    /// Origin plus size is used on both platforms so the meaning of each slot
+    /// is unambiguous wherever a snapshot is read.
+    pub(super) bounds: [i32; 4],
+    pub(super) screenshot_id: String,
+    // Pixel size of the delivered (possibly downscaled) screenshot. Model
+    // coordinates are expressed in this space and mapped back to physical
+    // window pixels before input is injected.
+    pub(super) display_width: u32,
+    pub(super) display_height: u32,
+}
+
+pub(super) struct AccessibilitySnapshot {
+    pub(super) window_hwnd: isize,
+    pub(super) elements: HashMap<String, NativeElement>,
+}
+
+#[derive(Default)]
+pub(super) struct ServerState {
+    pub(super) snapshots: HashMap<String, Snapshot>,
+    pub(super) accessibility: HashMap<String, AccessibilitySnapshot>,
+    pub(super) stopped_turn: Option<String>,
+}
+
+/// Bound document text by character count, flagging that more remains.
+///
+/// Counting characters rather than bytes keeps multi-byte text intact.
+pub(super) fn truncate_document_text(text: String) -> String {
+    if text.chars().count() <= DOC_TEXT_MAX {
+        return text;
+    }
+    let mut bounded: String = text.chars().take(DOC_TEXT_MAX).collect();
+    bounded.push_str("… (truncated)");
+    bounded
+}
+
+/// Render one accessibility element as a single line.
+///
+/// The format is part of the observation contract the skill documents, so it
+/// lives here rather than in either platform leaf.
+pub(super) fn element_line(element_id: &str, control_type_name: &str, name: &str) -> String {
+    format!("{element_id} {control_type_name} \"{name}\"")
+}
+
+/// An application discovered on disk but not currently showing a window.
+pub(super) struct InstalledApp {
+    pub(super) app_id: String,
+    pub(super) display_name: String,
+}
+
+/// Build the `list_apps` payload from installed applications and open windows.
+///
+/// An application that owns a window is reported as running and carries those
+/// windows; one found only on disk is reported with no windows. Both platforms
+/// share this so `is_running` can never mean different things on each.
+pub(super) fn merge_app_list(
+    installed: Vec<InstalledApp>,
+    windows: Vec<WindowInfo>,
+) -> Vec<Value> {
+    let mut order: Vec<String> = Vec::new();
+    let mut entries: HashMap<String, (String, bool, Vec<Value>)> = HashMap::new();
+    for app in installed {
+        if !entries.contains_key(&app.app_id) {
+            order.push(app.app_id.clone());
+        }
+        entries
+            .entry(app.app_id)
+            .or_insert((app.display_name, false, Vec::new()));
+    }
+    for window in windows {
+        let entry = entries.entry(window.app_id.clone()).or_insert_with(|| {
+            order.push(window.app_id.clone());
+            (window.display_name.clone(), false, Vec::new())
+        });
+        // A window proves the application is running, and its own display name
+        // is the one the user currently sees.
+        entry.0 = window.display_name.clone();
+        entry.1 = true;
+        entry.2.push(window.to_json());
+    }
+    order
+        .into_iter()
+        .filter_map(|app_id| {
+            let (display_name, is_running, windows) = entries.remove(&app_id)?;
+            Some(json!({
+                "id": app_id,
+                "display_name": display_name,
+                "is_running": is_running,
+                "windows": windows,
+            }))
+        })
+        .collect()
+}
+
+/// Map a coordinate expressed in screenshot space onto the window.
+///
+/// Returns the offset from the window's origin in physical pixels. Both
+/// platforms share the bounds check so a coordinate outside the delivered
+/// screenshot can never be extrapolated onto another application's window.
+pub(super) fn map_point(
+    snapshot: &Snapshot,
+    x: i64,
+    y: i64,
+) -> Result<(f64, f64), (&'static str, String)> {
+    let display_width = i64::from(snapshot.display_width.max(1));
+    let display_height = i64::from(snapshot.display_height.max(1));
+    if x < 0 || y < 0 || x >= display_width || y >= display_height {
+        return Err((
+            "point_outside_viewport",
+            "Point is outside the captured viewport.".to_string(),
+        ));
+    }
+    // The screenshot may have been downscaled, so scale back to the window's
+    // own pixels. With no downscaling these ratios are 1:1.
+    let width = f64::from(snapshot.bounds[2]);
+    let height = f64::from(snapshot.bounds[3]);
+    Ok((
+        x as f64 * width / display_width as f64,
+        y as f64 * height / display_height as f64,
+    ))
+}
+
+pub(super) fn next_id(prefix: &str) -> String {
+    format!("{prefix}-{}", NEXT_ID.fetch_add(1, Ordering::Relaxed))
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    fn snapshot(bounds: [i32; 4], display: (u32, u32)) -> Snapshot {
+        Snapshot {
+            window: WindowInfo {
+                hwnd: 1,
+                app_id: "app:test".to_string(),
+                display_name: "Test".to_string(),
+                title: String::new(),
+                class_name: String::new(),
+            },
+            bounds,
+            screenshot_id: "screenshot-1".to_string(),
+            display_width: display.0,
+            display_height: display.1,
+        }
+    }
+
+    fn window(app_id: &str, display_name: &str, hwnd: isize) -> WindowInfo {
+        WindowInfo {
+            hwnd,
+            app_id: app_id.to_string(),
+            display_name: display_name.to_string(),
+            title: String::new(),
+            class_name: String::new(),
+        }
+    }
+
+    #[test]
+    fn short_document_text_is_returned_unchanged() {
+        let text = "hello world".to_string();
+        assert_eq!(truncate_document_text(text.clone()), text);
+    }
+
+    #[test]
+    fn long_document_text_is_bounded_and_flagged() {
+        let bounded = truncate_document_text("x".repeat(DOC_TEXT_MAX + 500));
+        assert!(bounded.ends_with("… (truncated)"));
+        assert_eq!(
+            bounded.chars().filter(|value| *value == 'x').count(),
+            DOC_TEXT_MAX
+        );
+    }
+
+    #[test]
+    fn truncation_counts_characters_not_bytes() {
+        // Multi-byte text must not be cut mid-character.
+        let bounded = truncate_document_text("字".repeat(DOC_TEXT_MAX + 10));
+        assert_eq!(
+            bounded.chars().filter(|value| *value == '字').count(),
+            DOC_TEXT_MAX
+        );
+    }
+
+    #[test]
+    fn element_line_matches_the_listing_format() {
+        assert_eq!(
+            element_line("uia-1", "Edit", "text editor"),
+            "uia-1 Edit \"text editor\""
+        );
+    }
+
+    #[test]
+    fn a_point_inside_the_viewport_maps_by_proportion() {
+        // A 200x100 window delivered as a 100x50 screenshot is a 2:1 scale.
+        let snap = snapshot([10, 20, 200, 100], (100, 50));
+        let (x, y) = map_point(&snap, 50, 25).unwrap();
+        assert_eq!((x as i32, y as i32), (100, 50));
+    }
+
+    #[test]
+    fn an_unscaled_screenshot_maps_one_to_one() {
+        let snap = snapshot([0, 0, 100, 100], (100, 100));
+        let (x, y) = map_point(&snap, 30, 40).unwrap();
+        assert_eq!((x as i32, y as i32), (30, 40));
+    }
+
+    #[test]
+    fn points_outside_the_viewport_are_refused() {
+        let snap = snapshot([0, 0, 100, 100], (100, 100));
+        for (x, y) in [(-1, 0), (0, -1), (100, 0), (0, 100)] {
+            let error = map_point(&snap, x, y).expect_err("must be refused");
+            assert_eq!(error.0, "point_outside_viewport");
+        }
+    }
+
+    #[test]
+    fn a_running_application_reports_its_windows() {
+        let apps = merge_app_list(
+            Vec::new(),
+            vec![window("app:editor", "Editor", 1), window("app:editor", "Editor", 2)],
+        );
+        assert_eq!(apps.len(), 1);
+        assert_eq!(apps[0]["is_running"], serde_json::json!(true));
+        assert_eq!(apps[0]["windows"].as_array().unwrap().len(), 2);
+    }
+
+    #[test]
+    fn an_installed_application_reports_no_windows() {
+        let installed = vec![InstalledApp {
+            app_id: "app:/applications/notes.app".to_string(),
+            display_name: "Notes".to_string(),
+        }];
+        let apps = merge_app_list(installed, Vec::new());
+        assert_eq!(apps.len(), 1);
+        assert_eq!(apps[0]["is_running"], serde_json::json!(false));
+        assert!(apps[0]["windows"].as_array().unwrap().is_empty());
+    }
+
+    #[test]
+    fn a_running_application_is_not_duplicated_by_its_installed_entry() {
+        let installed = vec![InstalledApp {
+            app_id: "app:/applications/editor.app".to_string(),
+            // A stale on-disk name must not win over the live window's name.
+            display_name: "Editor 1.0".to_string(),
+        }];
+        let apps = merge_app_list(
+            installed,
+            vec![window("app:/applications/editor.app", "Editor", 1)],
+        );
+        assert_eq!(apps.len(), 1);
+        assert_eq!(apps[0]["is_running"], serde_json::json!(true));
+        assert_eq!(apps[0]["display_name"], serde_json::json!("Editor"));
+        assert_eq!(apps[0]["windows"].as_array().unwrap().len(), 1);
+    }
+}
